+++ tests/src/chanjo2/meta/test_handle_report_contents.py
@@ -3,8 +3,30 @@ from typing import List, Tuple
 from sqlalchemy.orm import sessionmaker
 
 from chanjo2.demo import DEMO_COVERAGE_QUERY_DATA
-from chanjo2.meta.handle_report_contents import get_missing_genes_from_db
+from chanjo2.meta.handle_report_contents import (
+    get_missing_genes_from_db,
+    get_report_data,
+)
 from chanjo2.models import SQLGene
+from chanjo2.models.pydantic_models import ReportQuery
+
+REPORT_EXPECTED_KEYS = [
+    "levels",
+    "extras",
+    "completeness_rows",
+    "incomplete_coverage_rows",
+    "default_level_completeness_rows",
+]
+REPORT_EXPECTED_EXTRA_KEYS = [
+    "panel_name",
+    "default_level",
+    "interval_type",
+    "case_name",
+    "hgnc_gene_ids",
+    "build",
+    "completeness_thresholds",
+    "samples",
+]
 
 
 def test_get_missing_genes_from_db(
@@ -21,3 +43,18 @@ def test_get_missing_genes_from_db(
     # THEN the get_missing_genes_from_db function should return the expected error, containing description and missing IDs
     assert missing_gene_error[0] == "Gene IDs not found in the database"
     assert missing_gene_error[1]
+
+
+def test_get_report_data(demo_session: sessionmaker):
+    """Test the function that collects the deta required for creating a coverage/overview report."""
+
+    # GIVEN a user query containing the expected parameters
+    query = ReportQuery.parse_obj(DEMO_COVERAGE_QUERY_DATA)
+    report_data: dict = get_report_data(query=query, session=demo_session)
+
+    # THEN get_report_data should return a dictionary with the expected report info
+    for expected_key in REPORT_EXPECTED_KEYS:
+        assert expected_key in report_data
+
+    for expected_key in REPORT_EXPECTED_EXTRA_KEYS:
+        assert expected_key in report_data["extras"]